[TASK] Move old partials from Components to Presets
Since TYPO3 uses the Components directory for "Fluid Components", all partials that are used for variants (e.g. in navigation) had to be moved to a separate directory. Since there are also SCSS and JS examples in the folders, the partials were not moved to the existing template folders. This has the advantage that later on, the FE Build can automatically search the folders for matching JS/SCSS and the path can be resolved with @sitekit-presets.